Added fetching of price info from IsThereAnyDeal in the AppDetailViewModel.kt

+++ app/src/main/java/com/github/khanshoaib3/steamcompanion/data/model/detail/isThereAnyDeal/PriceInfoResponse.kt
@@ -19,7 +19,7 @@ data class HistoryLow(
 
 @Serializable
 data class Price(
-    val amount: Double,
+    val amount: Float,
     val amountInt: Int,
     val currency: String
 )
